Rename update subcommand to upgrade
